address some CR comments
