+++ src/connection.ts
@@ -118,14 +118,13 @@ export default class Connection {
 		this.connection.onRequest(rt.WorkspaceSymbolsRequest.type, (params: rt.WorkspaceSymbolParamsWithLimit): Promise<SymbolInformation[]> => {
 			const enter = new Date().getTime();
 			return new Promise<SymbolInformation[]>((resolve, reject) => {
-				let result = [];
 				const init = new Date().getTime();
 				try {
 					return service.getWorkspaceSymbols(params.query, params.limit).then((result) => {
 						result = result ? result : [];
 						const exit = new Date().getTime();
 						console.error('workspace/symbol', params.query, 'total', (exit - enter) / 1000.0, 'busy', (exit - init) / 1000.0, 'wait', (init - enter) / 1000.0);
-						return resolve(result);
+						return resolve(result || []);
 					});
 				} catch (e) {
 					console.error(params, e);
@@ -137,14 +136,13 @@ export default class Connection {
 		this.connection.onRequest(rt.DocumentSymbolRequest.type, (params: DocumentSymbolParams): Promise<SymbolInformation[]> => {
 			const enter = new Date().getTime();
 			return new Promise<SymbolInformation[]>((resolve, reject) => {
-				let result = [];
 				const init = new Date().getTime();
 				try {
 					return service.getDocumentSymbol(params).then((result) => {
 						result = result ? result : [];
 						const exit = new Date().getTime();
 						console.error('textDocument/documentSymbol', "", 'total', (exit - enter) / 1000.0, 'busy', (exit - init) / 1000.0, 'wait', (init - enter) / 1000.0);
-						return resolve(result);
+						return resolve(result || []);
 					});
 				} catch (e) {
 					console.error(params, e);
@@ -153,9 +151,9 @@ export default class Connection {
 			});
 		});
 
-		this.connection.onRequest(rt.WorkspaceReferenceRequest.type, (params: util.WorkspaceReferenceParams): Promise<util.ReferenceInformation[]> => {
+		this.connection.onRequest(rt.WorkspaceReferenceRequest.type, (params: rt.WorkspaceReferenceParams): Promise<rt.ReferenceInformation[]> => {
 			const enter = new Date().getTime();
-			return new Promise<util.ReferenceInformation[]>((resolve, reject) => {
+			return new Promise<rt.ReferenceInformation[]>((resolve, reject) => {
 				let result = [];
 				const init = new Date().getTime();
 				try {
